Compress the (atom) type list
Adding an type list that includes all 118 elements of the periodic
table causes TensorFlow to run out of memory. So I am having to
compress the list to include only those elements that are actually
present.

This also has consequences for the DeePMD component where we will
now need to merge the type_map.raw maps across different chemical
systems.