Move molecule because of ASE peculiarities
The molecule needs to live in the simulation box. As there are
no periodic boundary conditions for this system LAMMPS will
simply ignore atoms that fall outside the box (and then LAMMPS
will complain that it didn't find all atoms and crash).

LAMMPS allows you to define a rectangular box anyway you want. ASE,
however, will generate a box in the positive quadrant with one
corner at the origin. The position furthest from the origin
is defined by taking the maximum of the atomic coordinates for
each dimension and multiplying it by 2 (and maybe adding some
padding). Of course NWChem, by default, will center the molecule
on the origin. The interplay between these 3 codes, therefore,
by default breaks everything.

To fix the problems I had originally, I needed to:
- shift the molecule so all atomic coordinates are positive
- shift the molecule so that by doubling the largest coordinates
  I get a simulation box of a suitable size.