Save current implementation, will change fundamentally.
The way ASE sets LAMMPS calculations up does not allow me to
generate the right files for DeepDriveMD. ASE does not let
me specify DCD as a LAMMPS trajectory format. MDAnalisys cannot
read a LAMMPS text trajectory. MDAnalysis also does not seem
to have an easy way to construct AtomGroup objects so I can
add structures from an arbitrary source.
The best opportunity would be to use ASE to read the LAMMPS text
trajectory and write it out as a sequence of XYZ structures. Then
read the XYZ trajectory with MDAnalysis and write it out as an DCD
file. Then I would have to read the DCD file and calculate the
contact maps and write those out again in HDF5. However, this
seems like a crazy amount of IO if I should just be able to tell
LAMMPS to write a DCD file.
So the next step is to remove ASE from the equation, and just
"hand write" the LAMMPS input file, like the way we do that for
NWChem.