Updates for different iterations
main1_nwchem.py generates a list of input files. The filenames (without
extension) are stored 'inputs.txt'.

If 'inputs.txt' does not exist then this is the first ab-initio stage
and we generate some initial input files.

If 'inputs.txt' exists already then we get here after a LAMMPS
simulation, and we generate new input files from the PDB files that
LAMMPS generated (and stored in a file).