Change the DFT solver
I have removed the spin unrestricted solver and replaced it with
the closed shell solver. The reason is that the molecule should not
be torn apart in the MD to the extend that we form radical species.
The practical aspect of this is that in order for the calculation
to produce radicals the spin symmetry has to break. This symmetry
breaking is unreliable as it depends on different numerical errors
accumulating for the alpha and beta electron wavefunctions. So it
typically takes a lot of iterations and it may or may not happen,
and therefore it is not clear that it produces a consistent data
set for the ML.

I also replaced the regular closed shell DFT solver with the
semi-Newton-Raphson solver. The latter much more reliably lowers
the energy (as required in an energy minization) as opposed to the
DIIS solver which often displays oscillatory behavior (especially
for strongly distorted geometries).