Adding minimization and fix list of PDB files
The list of PDB files was saved in the wrong format messing up the
routines that read that information.

The introduction of the energy minimization helps improve the starting
point for the MD simulations. If you start with an elevated temperature
at a point where the potential energy is not near a minimum then the
atoms may be accelerated at an unphysical rate causing weird things to
happen during the simulation. This then causes problems in the ab-initio
calculations. I also made the timestep shorter to avoid excessive
temperature excursions for relatively poorly optimized force fields.
For testing purposes we do not want to spend a tremendous amount of
time training the force field but then we also get relatively poorly
optimized force fields. Using short timesteps we can tolerate such
force fields and not get rediculous results.