Extent time limit and force switching between loops
Another 30 minute time limit caused the workflow to fail. Now
the time limit is set to 3000 minutes which at 50 hours is longer
than most batch queues allow for.

As we do not really want to wait for the workflow to "converge"
naturally I have added a file based mechanism to switch from the
force field training loop to the DeepDriveMD loop at a given point.
This mechanism should be removed before real applications.