DeGPUize the pytorch training
Running Pytorch on the GPU (A100 with 80GB of memory) ran out of memory
on the GPU in the backwards pass of autograd during training. To move
forward with the project I an therefore turning the GPU capabilities
off. The relevant changes are marked with "#HVD" so you can find them
easily.