N2P2 tool for symmetry function generation
This tool stems from the N2P2 project. It is discussed in the pull
request https://github.com/CompPhysVienna/n2p2/pull/15 but not
integrated in the main branch yet. So in the interim I am copying
it so we can use it here. Hopefully it will soon be integrated and
then we can use it from the "official" N2P2 distribution.